backend/app/core/security.py
from datetime import datetime, timedelta
from typing import Optional
from jose import JWTError, jwt
from passlib.context import CryptContext
from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from sqlalchemy.orm import Session
from .config import settings
from .database import get_db
from ..models.user import User

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
# HTTPBearer scheme for JSON-based login (extracts Bearer token from Authorization header)
bearer_scheme = HTTPBearer()

# Xác thực mật khẩu người dùng
def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)
# Hash mật khẩu người dùng
def get_password_hash(password):
    return pwd_context.hash(password)
# Lấy người dùng theo email
def get_user(db: Session, email: str):
    return db.query(User).filter(User.email == email).first()
# Lấy người dùng theo owner_id
def get_user_by_id(db: Session, owner_id: int):
    return db.query(User).filter(User.owner_id == owner_id).first()
# Xác thực người dùng
def authenticate_user(db: Session, email: str, password: str):
    user = get_user(db, email)
    if not user:
        return False
    if not verify_password(password, user.password):
        return False
    return user
# Tạo access token JWT
def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=settings.access_token_expire_minutes)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, settings.secret_key, algorithm=settings.algorithm)
    return encoded_jwt

# Lấy người dùng hiện tại từ token
async def get_current_user(credentials: HTTPAuthorizationCredentials = Depends(bearer_scheme), db: Session = Depends(get_db)):
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    token = credentials.credentials
    try:
        payload = jwt.decode(token, settings.secret_key, algorithms=[settings.algorithm])
        email: Optional[str] = payload.get("sub")
        owner_id: Optional[int] = payload.get("oid")
        if owner_id is not None:
            user = get_user_by_id(db, owner_id=owner_id)
        elif email is not None:
            user = get_user(db, email=email)
        else:
            raise credentials_exception
    except JWTError:
        raise credentials_exception
    if user is None:
        raise credentials_exception
    return user

# Lấy người dùng hiện tại và kiểm tra trạng thái hoạt động
async def get_current_active_user(current_user: User = Depends(get_current_user)):
    if not current_user.is_active:
        raise HTTPException(status_code=400, detail="Inactive user")
    return current_user

def require_role(required_role: str):
    """Decorator to check if user has required role"""
    async def role_checker(current_user: User = Depends(get_current_active_user), db: Session = Depends(get_db)):
        # Load role relationship if not already loaded
        if not hasattr(current_user, 'role') or current_user.role is None:
            db.refresh(current_user, ['role'])
        
        if current_user.role.authority != required_role:
            raise HTTPException(
                status_code=status.HTTP_403_FORBIDDEN,
                detail=f"Access denied. Required role: {required_role}"
            )
        return current_user
    return role_checker
