backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
python-dotenv
alembic
google-generativeai
httpx
